update shell for exception.. PID file..

+++ stop-stage.sh
@@ -1,3 +1,14 @@
+#/bin/bash
+
 cd target/universal/stage
-kill -9 `cat RUNNING_PID`                
-rm RUNNING_PID 
+
+if [[ -f RUNNING_PID ]] && ps -p $(cat RUNNING_PID) > /dev/null 2>&1
+then
+   kill $(cat RUNNING_PID)
+   rm RUNNING_PID
+   echo "Stage Stopped.."
+else
+   echo "Not Running...."
+fi
+
+exit 0
